✨ Fix logging; add auto-reload; add power units
The logging code was broken (I don't know why/how I thought it would work).
I'm not sure that there is any difference between console.log, console.info,
and other logging functions. The new code should handle log levels better.

Also, there is a huge memory leak somewhere. I don't know if it's in
my code on in the zwave library or somewhere else, but it leads to OOM
errors in just a few hours. I've added a routine that terminates the
program if memory usage (measuring RSS) goes above a given threshold.
On my system, it means that the exporter restarts every hour or so,
but that's still better than crashing (or worse, swapping and killing
a SDcard in a few hours).

Finally, since I'm now using this to measure power usage, I've added
units like A, V, W...